src/cli/commands/cluster.py
"""Clustering command implementation."""

from pathlib import Path
from typing import Dict, Any
import pickle
import numpy as np
from sklearn.cluster import KMeans, GaussianMixture, SpectralClustering


def cluster_codes(
    codes_file: Path,
    method: str,
    n_clusters: int,
    min_support: float
) -> Dict[str, Any]:
    """Cluster FSQ codes into behavioral motifs."""
    # Load codes
    with open(codes_file, 'rb') as f:
        data = pickle.load(f)
    
    codes = data['codes']
    
    # Perform clustering
    if method == "kmeans":
        clusterer = KMeans(n_clusters=n_clusters, random_state=42)
    elif method == "gmm":
        clusterer = GaussianMixture(n_components=n_clusters, random_state=42)
    elif method == "spectral":
        clusterer = SpectralClustering(n_clusters=n_clusters, random_state=42)
    
    # For stub, simulate clustering
    labels = np.random.randint(0, n_clusters, size=len(codes))
    
    # Calculate durations
    durations = {}
    for i in range(n_clusters):
        mask = labels == i
        if mask.sum() > 0:
            # Simulate duration calculation
            durations[i] = np.random.uniform(20, 50)
    
    return {
        'labels': labels,
        'n_clusters': n_clusters,
        'durations': durations,
        'method': method
    }


def save_clusters(output_dir: Path, results: Dict[str, Any]):
    """Save clustering results."""
    output_dir.mkdir(parents=True, exist_ok=True)
    
    with open(output_dir / "clusters.pkl", 'wb') as f:
        pickle.dump(results, f)
    
    np.save(output_dir / "labels.npy", results['labels'])